Update: route endpoints to support new output objects. Reformatted code.

+++ social-routing-service/src/main/kotlin/ps/g49/socialroutingservice/models/domainModel/Route.kt
@@ -8,7 +8,7 @@ data class Route(
         val name: String,
         val description: String? = null,
         var rating: Double? = null, // created by db
-        var duration: Long? = null,// TODO created by service
+        var duration: Int? = null,// TODO created by service
         var dateCreated: Date? = null, // created by db
         val points: List<Point>? = null,
         val personIdentifier: Int
